requirements.txt
alembic
asyncpg
fastapi
gunicorn
psycopg2
python-dotenv
sqlalchemy --pre
uvicorn
uvloop
httptools
